# docs/swagger.yaml
basePath: /
definitions:
...
  book.CreateBookDTO:
    description: Данные для создания новой книги
    properties:
      author:
        description: |-
          @Description Автор книги
          @example F. Scott Fitzgerald
        maxLength: 100
        minLength: 3
        type: string
      description:
        description: |-
          @Description Описание книги
          @example A story of the fabulously wealthy Jay Gatsby and his love for the beautiful Daisy Buchanan.
        maxLength: 1000
        minLength: 10
        type: string
      photos:
        description: |-
          @Description Фотографии книги в формате base64
          @example ["data:image/jpeg;base64,/9j/4AAQSkZJRg..."]
        items:
          type: string
        maxItems: 5
        minItems: 1
        type: array
      state_id:
        description: |-
          @Description ID состояния книги
          @example 1
        type: integer
      tag_ids:
        description: |-
          @Description ID тегов книги
          @example [1, 2, 3]
        items:
          type: integer
        maxItems: 5
        minItems: 1
        type: array
      title:
        description: |-
          @Description Название книги
          @example The Great Gatsby
        maxLength: 100
        minLength: 3
        type: string
    required:
    - author
    - description
    - photos
    - state_id
    - tag_ids
    - title
    type: object
...
  user.UpdateUserDTO:
    properties:
      avatar:
        type: string
      description:
        maxLength: 1000
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API for book trading system with tag support
  title: Book Trading API
  version: "1.0"
paths:
  /api/v1/books:
    get:
      description: Get a list of all books
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/book.Book'
            type: array
      summary: Get all books
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Create a new book with the given details
      parameters:
      - description: Book details
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/book.CreateBookDTO'
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/book.Book'
      security:
      - BearerAuth: []
      summary: Create a new book
      tags:
      - books
  /api/v1/books/{id}:
    delete:
      description: Delete a book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete book
      tags:
      - books
    get:
      description: Get book information by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/book.Book'
      summary: Get book by ID
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Update book details
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Book details
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/book.UpdateBookDTO'
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/book.Book'
      security:
      - BearerAuth: []
      summary: Update a book
      tags:
      - books
  /api/v1/books/{id}/state:
    patch:
      consumes:
      - application/json
      description: Update the state of a book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: New state
        in: body
        name: state
        required: true
        schema:
          $ref: '#/definitions/book.UpdateBookStateDTO'
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/book.Book'
      security:
      - BearerAuth: []
      summary: Update book state
      tags:
      - books
  /api/v1/books/{id}/tags:
    post:
      consumes:
      - application/json
      description: Add tags to an existing book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tag IDs
        in: body
        name: tagIds
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/book.Book'
      security:
      - BearerAuth: []
      summary: Add tags to book
      tags:
      - books
  /api/v1/books/search:
    get:
      description: Search books by tag IDs
      parameters:
      - collectionFormat: csv
        description: Tag IDs
        in: query
        items:
          type: integer
        name: tagIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/book.Book'
            type: array
      summary: Search books by tags
      tags:
      - books
  /api/v1/states:
    get:
      description: Get list of all book states
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/state.State'
            type: array
      summary: Get all states
      tags:
      - states
    post:
      consumes:
      - application/json
      description: Create a new book state
      parameters:
      - description: State object
        in: body
        name: state
        required: true
        schema:
          $ref: '#/definitions/state.CreateStateDTO'
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/state.State'
      security:
      - BearerAuth: []
      summary: Create a new state
      tags:
      - states
  /api/v1/states/{id}:
    delete:
      description: Delete a book state
      parameters:
      - description: State ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete state
      tags:
      - states
    get:
      description: Get a book state by ID
      parameters:
      - description: State ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/state.State'
      summary: Get state by ID
      tags:
      - states
    put:
      consumes:
      - application/json
      description: Update a book state
      parameters:
      - description: State ID
        in: path
        name: id
        required: true
        type: integer
      - description: State object
        in: body
        name: state
        required: true
        schema:
          $ref: '#/definitions/state.UpdateStateDTO'
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/state.State'
      security:
      - BearerAuth: []
      summary: Update state
      tags:
      - states
  /api/v1/tags:
    get:
      description: Get list of all tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tag.Tag'
            type: array
      summary: Get all tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: Create a new tag with the given name
      parameters:
      - description: Tag details
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/tag.CreateTagDTO'
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.Tag'
      security:
      - BearerAuth: []
      summary: Create a new tag
      tags:
      - tags
  /api/v1/tags/{id}:
    delete:
      description: Delete a tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete tag
      tags:
      - tags
    get:
      description: Get tag information by ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.Tag'
      summary: Get tag by ID
      tags:
      - tags
  /api/v1/tags/popular:
    get:
      description: Get list of popular tags
      parameters:
      - description: Number of tags to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tag.Tag'
            type: array
      summary: Get popular tags
      tags:
      - tags
  /api/v1/users:
    get:
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      description: Delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Get user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /api/v1/users/{id}/books:
    get:
      description: Get paginated list of user's books
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: pageSize
        type: integer
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns books and pagination info
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user books
      tags:
      - books
...
swagger: "2.0"